fix(console): draging anchor in the color picker on application branding page
