+++ packages/console/package.json
@@ -93,7 +93,7 @@
     "property-information": "^6.2.0",
     "react": "^18.3.1",
     "react-animate-height": "^3.0.4",
-    "react-color": "^2.19.3",
+    "react-color-palette": "^7.2.1",
     "react-confetti": "^6.1.0",
     "react-dnd": "^16.0.0",
     "react-dnd-html5-backend": "^16.0.0",
