+++ packages/console/src/ds-components/ColorPicker/index.module.scss
@@ -28,3 +28,7 @@
     margin-right: _.unit(2);
   }
 }
+
+.palette {
+  width: 400px;
+}
